+++ src/main/kotlin/com/github/copyinaction/coupon/domain/DiscountMethod.kt
@@ -1,6 +1,0 @@
-package com.github.copyinaction.coupon.domain
-
-enum class DiscountMethod(val description: String) {
-    FIXED("정액 할인"),
-    PERCENTAGE("정률 할인")
-}